Refactoring code and updates dependencies
