--- client/.gitignore
@@ -1,3 +1,0 @@
-out
-server
-node_modules
