--- client/tsconfig.json
@@ -1,14 +1,0 @@
-{
-	"compilerOptions": {
-		"target": "ES5",
-		"module": "commonjs",
-		"moduleResolution": "node",
-		"outDir": "out",
-		"noLib": true,
-		"sourceMap": true
-	},
-	"exclude": [
-		"node_modules",
-		"server"
-	]
-}
