+++ package.json
@@ -2,7 +2,7 @@
 	"name": "javac-linter",
 	"displayName": "Java Linter",
 	"description": "Java Linter for VS Code",
-	"version": "1.2.6",
+	"version": "1.3.0",
 	"icon": "images/icon.png",
 	"publisher": "faustinoaq",
 	"license": "MIT",
@@ -30,7 +30,7 @@
 	"activationEvents": [
 		"onLanguage:java"
 	],
-	"main": "./out/src/extension",
+	"main": "./out/src/Main",
 	"contributes": {
 		"configuration": {
 			"type": "object",
@@ -48,6 +48,9 @@
 				},
 				"javac-linter.classpath": {
 					"type": "array",
+					"default": [
+						"."
+					],
 					"items": {
 						"type": "string"
 					},
@@ -62,15 +65,15 @@
 		}
 	},
 	"scripts": {
-		"vscode:prepublish": "node ./node_modules/vscode/bin/compile",
-		"compile": "node ./node_modules/vscode/bin/compile -watch -p ./",
+		"vscode:prepublish": "tsc -p ./",
+		"compile": "tsc -watch -p ./",
 		"postinstall": "node ./node_modules/vscode/bin/install"
 	},
 	"devDependencies": {
-		"typescript": "^1.8.9",
-		"vscode": "^0.11.0"
-	},
-	"dependencies": {
-		"vscode-languageclient": "^2.2.1"
+		"typescript": "^2.7.2",
+		"vscode": "^1.1.10",
+		"mocha": "^3.5.0",
+		"@types/node": "^8.0.26",
+		"@types/mocha": "^2.2.42"
 	}
 }
