--- server/.gitignore
@@ -1,2 +1,0 @@
-out
-node_modules
