--- server/README.md
@@ -1,3 +1,0 @@
-# README
-
-A language server that check Java code with `javac`
