--- server/package.json
@@ -1,20 +1,0 @@
-{
-	"name": "javac-linter-server",
-	"description": "Language server in node using exec javac",
-	"version": "1.2.6",
-	"author": "faustinoaq",
-	"license": "MIT",
-	"engines": {
-		"node": "*"
-	},
-	"dependencies": {
-		"vscode-languageserver": "^2.2.0"
-	},
-	"devDependencies": {
-		"typescript": "^2.0.3"
-	},
-	"scripts": {
-		"compile": "installServerIntoExtension ../client ./package.json ./tsconfig.json && tsc -p .",
-		"watch": "installServerIntoExtension ../client ./package.json ./tsconfig.json && tsc --watch -p ."
-	}
-}
