--- server/tsconfig.json
@@ -1,12 +1,0 @@
-{
-	"compilerOptions": {
-		"target": "ES5",
-		"module": "commonjs",
-		"moduleResolution": "node",
-		"sourceMap": true,
-		"outDir": "../client/server"
-	},
-	"exclude": [
-		"node_modules"
-	]
-}
